+++ include/func.hpp
@@ -20,4 +20,4 @@ struct StopWatch
 
 void lineRender(SDL_Renderer* aRenderer, std::vector<bool>& aMandel, int& aDisplacement);
 
-void mandelDraw ( SDL_Renderer* aRenderer, const int aScreenWidth, const int aScreenHeight, const float numWidth);
+void mandelDraw ( SDL_Renderer* aRenderer, const int aScreenWidth, const int aScreenHeight, const double numWidth);
